chore(marketing-analytics): extract session breakdown query runner base
The two attribution runners each carried their own copy of the session
dimension plumbing. Anything else that wants to slice sessions the same way had
to import an attribution runner or restate the rules and risk drifting.

Moves the breakdown expression, the source and campaign normalization, the
pageview and touchpoint conditions, and the conversion goal resolution onto a
shared base. No behavior changes.

Generated-By: PostHog Desktop
Task-Id: f3d141cb-c6a6-47c8-b20e-7a4a55d864c7
